+++ zeroeq/connection/broker.h
@@ -90,7 +90,7 @@ private:
 
     // Receiver API
     void addSockets( std::vector< zeroeq::detail::Socket >& entries ) final;
-    void process( zeroeq::detail::Socket& socket ) final;
+    void process( zeroeq::detail::Socket& socket, uint32_t timeout ) final;
     void addConnection( const std::string& ) final { ZEROEQDONTCALL; } // LCOV_EXCL_LINE
 };